Fix prefetching to actually store prefetched data in the objects.
In query-variable-binder2, always make use of the prefetched slots for
query variables used as actions ("MORE SPEED").

This applies to the following usual pattern:
  (prc:select (x)
    (prc:from (x x))
    ...)
which didn't fill slots previously.

(A few commits earlier, I fixed prefetching so that it would actually
prefetch columns from the database, which had been broken.  Now this
commits is needed so that perec also _uses_ the prefetched data to fill
the objects.)